+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 - Added Netlist search functionality to search by instance path
 - Added Zoom animations for highlight zoom (configurable in settings)
 - Added Cmd + Scroll to zoom on Mac OS
+- Added User settings for renderer
+  - Animations enabled and duration
+  - Disable Analog Renderer optimizations
+  - Override window.devicePixelRatio
+  - Multi bit - preserve height vs preserve shape
 
 # 1.5.0 - 2/25/2026 - Latest Release